[HttpKernel] Add `MapSessionParameter` to map session parameters to controllers
This attribute allow to read a value from the session.
The basic usage is using `#[MapSessionParameter] string $parameter` in a controller.

The name of the parameter is used as the session key and can be overriden by the name parameter of the attribute. It can be a string list, in this case, every key in the list is provided.
The close parameter allow to close the session after reading if the controller doesn't need to write in session. This can help if having the session open keep a lock preventing other request from running.
The defaultValue parameter is mainly here for variadic argument, if not provided, the default value of the parameter will be used, or null.

In case of variadic parameter, if the variable is not set in session and no default value is provided, nothing will be resolved.